add typed redux hooks, and typed store and state values
