+++ client/src/features/hooks.ts
@@ -1,0 +1,9 @@
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+import type { RootState, AppDispatch } from "./store";
+
+/**
+ * Typed versions of the useDispatch and useSelector hooks for usage in your application,
+ * instead of typing them every time we use these hooks
+ */
+export const useAppDispatch: () => AppDispatch = useDispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
